Pick up the latest SASS updates from the brand repo
This _only_ picks up the SASS changes, partly as I'm not sure
which way around changes tend to flow and partly as I'm not sure
what content here (if any) would need to change with these SASS
updates.

There might need to be a follow-on which adds some classes to the
HTML here.